+++ spec/models/results_spec.rb
